refactor: :recycle: extract func modulos
