+++ include/system_utils.h
@@ -1,0 +1,8 @@
+#ifndef system_utils_h
+#define system_utils_h
+
+void pause();
+void clean();
+
+
+#endif
